Move User Password Settings to Auth
