+++ components/ILIAS/User/classes/Settings/class.ilPersonalSettingsGUI.php
@@ -18,24 +18,20 @@
 
 declare(strict_types=1);
 
+use ILIAS\Authentication\Password\ilLocalPasswordSettingsGUI;
 use ILIAS\DI\LoggingServices;
-use ILIAS\UI\Component\Input\Container\Form\Standard as Form;
-use ILIAS\UI\Component\Input\Field\Password;
-use ILIAS\UI\Component\Input\Field\Section;
 use ILIAS\UI\Factory as UIFactory;
 use ILIAS\UI\Renderer;
-use ILIAS\User\UserGUIRequest;
-use Psr\Http\Message\ServerRequestInterface;
+use ILIAS\Authentication\Password\ilLocalPasswordManager;
 
 /**
  * GUI class for personal profile
  * @author       Alexander Killing <killing@leifos.de>
- * @ilCtrl_Calls ilPersonalSettingsGUI: ilMailOptionsGUI
+ * @ilCtrl_Calls ilPersonalSettingsGUI: ilMailOptionsGUI, ILIAS\Authentication\Password\ilLocalPasswordSettingsGUI
  */
 class ilPersonalSettingsGUI
 {
     private ilPropertyFormGUI $form;
-    private UserGUIRequest $request;
     private ilGlobalTemplateInterface $tpl;
     private UIFactory $ui_factory;
     private Renderer $ui_renderer;
@@ -80,11 +76,6 @@ class ilPersonalSettingsGUI
         $this->ctrl->saveParameter($this, 'user_page');
 
         $this->user_settings_config = new ilUserSettingsConfig();
-        $this->request = new UserGUIRequest(
-            $DIC->http(),
-            $DIC['refinery']
-        );
-
         $this->starting_point_repository = new ilUserStartingPointRepository(
             $this->user,
             $DIC['ilDB'],
@@ -116,6 +107,13 @@ class ilPersonalSettingsGUI
                 $this->ctrl->forwardCommand(new ilMailOptionsGUI());
 
                 break;
+            case strtolower(ilLocalPasswordSettingsGUI::class):
+                $this->initSubTabs('showPersonalData');
+                $this->tabs->activateTab('password');
+                $this->setHeader();
+                $this->ctrl->forwardCommand(new ilLocalPasswordSettingsGUI());
+
+                break;
             default:
                 $cmd = $this->ctrl->getCmd('showGeneralSettings');
                 $this->$cmd();
@@ -140,10 +138,16 @@ class ilPersonalSettingsGUI
             $showGeneralSettings
         );
 
-        if ($this->allowPasswordChange()) {
+        if (ilLocalPasswordManager::getInstance()->allowPasswordChange($this->user)) {
             $this->tabs->addTarget(
                 'password',
-                $this->ctrl->getLinkTarget($this, 'showPassword'),
+                $this->ctrl->getLinkTargetByClass(
+                    [
+                        ilDashboardGUI::class,
+                        self::class,
+                        ilLocalPasswordSettingsGUI::class],
+                    'showPassword'
+                ),
                 '',
                 '',
                 '',
@@ -176,205 +180,6 @@ class ilPersonalSettingsGUI
     public function setHeader(): void
     {
         $this->tpl->setTitle($this->lng->txt('personal_settings'));
-    }
-
-    public function showPassword(
-        Form $form = null,
-        bool $hide_form = false
-    ): void {
-        $this->initSubTabs('showPersonalData');
-        $this->tabs->activateTab('password');
-
-        $this->setHeader();
-        // check whether password of user have to be changed
-        // due to first login or password of user is expired
-        if ($this->user->isPasswordChangeDemanded()) {
-            $this->tpl->setOnScreenMessage('info', $this->lng->txt('password_change_on_first_login_demand'));
-        } elseif ($this->user->isPasswordExpired()) {
-            $msg = $this->lng->txt('password_expired');
-            $password_age = $this->user->getPasswordAge();
-            $this->tpl->setOnScreenMessage('info', sprintf($msg, $password_age));
-        }
-
-        if (!$form && !$hide_form) {
-            $form = $this->getPasswordForm();
-        }
-        $this->tpl->setContent(!$hide_form ? $this->ui_renderer->render($form) : '');
-        $this->tpl->printToStdout();
-    }
-
-    public function getPasswordForm(
-        ServerRequestInterface $request = null,
-        array $errors = []
-    ): Form {
-        $items = [];
-        if ($this->allowPasswordChange()) {
-            $pw_info_set = false;
-            if ((int) $this->user->getAuthMode(true) === ilAuthUtils::AUTH_LOCAL) {
-                $cpass = $this->ui_factory->input()->field()->password(
-                    $this->lng->txt('current_password'),
-                    ilSecuritySettingsChecker::getPasswordRequirementsInfo()
-                );
-
-                $pw_info_set = true;
-                if ($this->user->getPasswd()) {
-                    $cpass = $cpass->withRequired(true);
-                }
-                $cpass = $cpass->withRevelation(true);
-                $cpass_error = $errors['current_password'] ?? [];
-                if ($cpass_error !== []) {
-                    $cpass = $cpass->withError(implode('<br>', $cpass_error));
-                }
-
-                $items['current_password'] = $cpass;
-            }
-
-            // new password
-            $ipass = $this->ui_factory->input()->field()->password(
-                $this->lng->txt('desired_password'),
-            );
-            if ($pw_info_set === false) {
-                $ipass = $ipass->withByline(ilSecuritySettingsChecker::getPasswordRequirementsInfo());
-            }
-            $ipass = $ipass->withRequired(true);
-            $ipass = $ipass->withRevelation(true);
-            $ipass_error = $errors['new_password'] ?? [];
-            if ($ipass_error !== []) {
-                $ipass = $ipass->withError(implode('<br>', $ipass_error));
-            }
-            $items['new_password'] = $ipass;
-
-            switch ($this->user->getAuthMode(true)) {
-                case ilAuthUtils::AUTH_LOCAL:
-                    $title = $this->lng->txt('chg_password');
-
-                    break;
-                case ilAuthUtils::AUTH_SHIBBOLETH:
-                case ilAuthUtils::AUTH_CAS:
-                    if (ilDAVActivationChecker::_isActive()) {
-                        $title = $this->lng->txt('chg_ilias_and_webfolder_password');
-                    } else {
-                        $title = $this->lng->txt('chg_ilias_password');
-                    }
-
-                    break;
-                default:
-                    $title = $this->lng->txt('chg_ilias_password');
-
-                    break;
-            }
-            $section = $this->ui_factory->input()->field()->section($items, $title);
-            $items = ['password' => $section];
-        }
-
-        return $this->ui_factory->input()->container()->form()->standard(
-            $this->ctrl->getLinkTarget($this, 'savePassword'),
-            $items
-        )->withSubmitLabel($this->lng->txt('save'));
-    }
-
-    /**
-     * Check, whether password change is allowed for user
-     */
-    protected function allowPasswordChange(): bool
-    {
-        if (ilSession::get('used_external_auth_mode')) {
-            return false;
-        }
-
-        $status = ilAuthUtils::isPasswordModificationEnabled($this->user->getAuthMode(true));
-        if ($status) {
-            return true;
-        }
-
-        return ilAuthUtils::isPasswordModificationHidden() &&
-            ($this->user->isPasswordChangeDemanded() || $this->user->isPasswordExpired());
-    }
-
-    public function savePassword(): void
-    {
-        if (!$this->allowPasswordChange()) {
-            $this->ctrl->redirect($this, 'showPersonalData');
-
-            return;
-        }
-
-        $form = $this->getPasswordForm()->withRequest($this->request->getRequest());
-        $section = $form->getInputs()['password'];
-        /** @var Password $cp
-         * @var Password $np
-         */
-        $cp = $section->getInputs()['current_password'];
-        $np = $section->getInputs()['new_password'];
-        $errors = ['current_password' => [], 'new_password' => []];
-
-        $entered_current_password = $cp->getValue();
-        $entered_new_password = $np->getValue();
-
-        if (!$form->getError()) {
-            $data = $form->getData();
-            /** @var Section $section */
-            $error = false;
-
-            if ((int) $this->user->getAuthMode(true) === ilAuthUtils::AUTH_LOCAL) {
-                if (!ilUserPasswordManager::getInstance()->verifyPassword(
-                    $this->user,
-                    $entered_current_password
-                )) {
-                    $error = true;
-                    $errors['current_password'][] = $this->lng->txt('passwd_wrong');
-                }
-            }
-
-            if (!ilSecuritySettingsChecker::isPassword($entered_new_password, $custom_error)) {
-                $error = true;
-                if ($custom_error !== '') {
-                    $errors['new_password'][] = $custom_error;
-                } else {
-                    $errors['new_password'][] = $this->lng->txt('passwd_invalid');
-                }
-            }
-            $error_lng_var = '';
-            if (!ilSecuritySettingsChecker::isPasswordValidForUserContext(
-                $entered_new_password,
-                $this->user,
-                $error_lng_var
-            )) {
-                $this->tpl->setOnScreenMessage('failure', $this->lng->txt('form_input_not_valid'));
-                $errors['new_password'][] = $this->lng->txt($error_lng_var);
-                $error = true;
-            }
-            if (
-                $entered_current_password === $entered_new_password &&
-                ($this->user->isPasswordExpired() || $this->user->isPasswordChangeDemanded())
-            ) {
-                $error = true;
-                $errors['new_password'][] = $this->lng->txt('new_pass_equals_old_pass');
-            }
-
-            if (!$error) {
-                $this->user->resetPassword($entered_new_password, $entered_new_password);
-                if ($entered_current_password !== $entered_new_password) {
-                    $this->user->setLastPasswordChangeToNow();
-                    $this->user->setPasswordPolicyResetStatus(false);
-                    $this->user->update();
-                }
-
-                if (ilSession::get('orig_request_target')) {
-                    $this->tpl->setOnScreenMessage('success', $this->lng->txt('saved_successfully'), true);
-                    $target = ilSession::get('orig_request_target');
-                    ilSession::set('orig_request_target', '');
-                    ilUtil::redirect($target);
-                } else {
-                    $this->tpl->setOnScreenMessage('success', $this->lng->txt('saved_successfully'));
-                    $this->showPassword(null, true);
-
-                    return;
-                }
-            }
-        }
-
-        $this->showPassword($this->getPasswordForm($this->request->getRequest(), $errors));
     }
 
     public function workWithUserSetting(string $setting): bool
